+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/message/abstracts/PossiblySentViaBot.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.CommonBot
+
+interface PossiblySentViaBot {
+    val senderBot: CommonBot?
+}