+++ TelegramBotAPI/src/commonMain/kotlin/com/github/insanusmokrassar/TelegramBotAPI/utils/MimeType.kt
@@ -7,5 +7,7 @@ expect class MimeType {
     val raw: String
 }
 
+expect fun buildMimeType(raw: String): MimeType
+
 @Serializer(MimeType::class)
 internal expect object MimeTypeSerializer : KSerializer<MimeType>